refactor/move source static functions
